# refactor(machines): cache machine list selectors