--- a/package.json
+++ b/package.json
@@ -53,6 +53,7 @@
     "js-file-download": "0.4.12",
     "path-parse": "1.0.7",
     "pluralize": "8.0.0",
+    "re-reselect": "4.0.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-dropzone": "14.2.3",