--- a/src/app/base/components/TableMenu/TableMenu.tsx
+++ b/src/app/base/components/TableMenu/TableMenu.tsx
@@ -1,5 +1,10 @@
+import { useMemo } from "react";
+
+import type {
+  ContextualMenuProps,
+  MenuLink,
+} from "@canonical/react-components";
 import { ContextualMenu } from "@canonical/react-components";
-import type { ContextualMenuProps } from "@canonical/react-components";
 import classNames from "classnames";
 
 export type Props<L = null> = {
@@ -16,23 +21,31 @@ export type Props<L = null> = {
 const TableMenu = <L,>({
   className,
   disabled = false,
-  links,
+  // If there are no links then make it an empty array so that it can be validly spread below.
+  links = [],
   title,
   onToggleMenu,
   position = "left",
   positionNode,
   "aria-label": ariaLabel,
 }: Props<L>): JSX.Element => {
-  // If there are no links then make it an empty array so that it can be validly spread below.
-  links = links || [];
+  const linksWithTitle = useMemo(
+    () => [
+      ...(title ? [title] : []),
+      ...(Array.isArray(links) ? links : [links]),
+    ],
+    [title, links]
+  );
+  const toggleProps = useMemo(
+    () => ({ "aria-label": ariaLabel || title || undefined }),
+    [ariaLabel, title]
+  );
+
   return (
     <ContextualMenu
       className={classNames("p-table-menu", className)}
       hasToggleIcon
-      links={[
-        ...(title ? [title] : []),
-        ...(Array.isArray(links) ? links : [links]),
-      ]}
+      links={linksWithTitle as MenuLink<L | null>[]}
       onToggleMenu={onToggleMenu || undefined}
       position={position}
       positionNode={positionNode || undefined}
@@ -41,7 +54,7 @@ const TableMenu = <L,>({
       toggleAppearance="base"
       toggleClassName="u-no-margin--bottom p-table-menu__toggle"
       toggleDisabled={disabled || false}
-      toggleProps={{ "aria-label": ariaLabel || title || undefined }}
+      toggleProps={toggleProps}
     />
   );
 };