--- a/src/app/machines/views/MachineList/MachineListTable/OwnerColumn/OwnerColumn.tsx
+++ b/src/app/machines/views/MachineList/MachineListTable/OwnerColumn/OwnerColumn.tsx
@@ -1,10 +1,11 @@
-import { memo, useEffect, useState } from "react";
+import { memo, useCallback, useEffect, useMemo, useState } from "react";
 
 import { Spinner } from "@canonical/react-components";
 import { useSelector } from "react-redux";
 
 import DoubleRow from "app/base/components/DoubleRow";
 import { useMachineActions } from "app/base/hooks";
+import type { MachineMenuAction } from "app/base/hooks/node";
 import { useToggleMenu } from "app/machines/hooks";
 import machineSelectors from "app/store/machine/selectors";
 import type { Machine, MachineMeta } from "app/store/machine/types";
@@ -19,6 +20,8 @@ type Props = {
   systemId: Machine[MachineMeta.PK];
   showFullName?: boolean;
 };
+
+const actions: MachineMenuAction[] = [NodeActions.ACQUIRE, NodeActions.RELEASE];
 
 export const OwnerColumn = ({
   onToggleMenu,
@@ -41,15 +44,17 @@ export const OwnerColumn = ({
     : machine?.owner || "-";
   const tagsDisplay = getTagsDisplay(machineTags);
 
+  const handleMachineActionClick = useCallback(() => {
+    if (machine) {
+      setUpdating(machine.status);
+    }
+  }, [machine]);
+
   const menuLinks = useMachineActions(
     systemId,
-    [NodeActions.ACQUIRE, NodeActions.RELEASE],
+    actions,
     "No owner actions available",
-    () => {
-      if (machine) {
-        setUpdating(machine.status);
-      }
-    }
+    handleMachineActionClick
   );
 
   useEffect(() => {
@@ -58,25 +63,32 @@ export const OwnerColumn = ({
     }
   }, [updating, machine?.status]);
 
+  const primary = useMemo(
+    () => (
+      <>
+        {updating === null ? null : <Spinner className="u-nudge-left--small" />}
+        <span data-testid="owner">{ownerDisplay}</span>
+      </>
+    ),
+    [updating, ownerDisplay]
+  );
+  const secondary = useMemo(
+    () => (
+      <span data-testid="tags" title={tagsDisplay}>
+        {tagsDisplay}
+      </span>
+    ),
+    [tagsDisplay]
+  );
+
   return (
     <DoubleRow
       menuLinks={onToggleMenu ? menuLinks : null}
       menuTitle="Take action:"
       onToggleMenu={toggleMenu}
-      primary={
-        <>
-          {updating === null ? null : (
-            <Spinner className="u-nudge-left--small" />
-          )}
-          <span data-testid="owner">{ownerDisplay}</span>
-        </>
-      }
+      primary={primary}
       primaryTitle={ownerDisplay}
-      secondary={
-        <span data-testid="tags" title={tagsDisplay}>
-          {tagsDisplay}
-        </span>
-      }
+      secondary={secondary}
       secondaryTitle={tagsDisplay}
     />
   );