--- a/src/app/store/machine/selectors.ts
+++ b/src/app/store/machine/selectors.ts
@@ -1,5 +1,6 @@
 import type { Selector } from "@reduxjs/toolkit";
 import { createSelector } from "@reduxjs/toolkit";
+import { createCachedSelector } from "re-reselect";
 
 import { ACTIONS } from "app/store/machine/slice";
 import type {
@@ -529,15 +530,11 @@ const listLoading = createSelector(
  * @param selected - Whether to filter for selected machines.
  * @returns A list of machines.
  */
-const list = createSelector(
-  [
-    machineState,
-    defaultSelectors.all,
-    (_state: RootState, callId: string | null | undefined) => ({
-      callId,
-    }),
-  ],
-  (machineState, allMachines, { callId }) => {
+const list = createCachedSelector(
+  machineState,
+  defaultSelectors.all,
+  (_state: RootState, callId: string | null | undefined) => callId,
+  (machineState, allMachines, callId) => {
     const machines: Machine[] = [];
     getList(machineState, callId)?.groups?.forEach((group) => {
       group.items.forEach((systemId) => {
@@ -551,7 +548,7 @@ const list = createSelector(
     });
     return machines;
   }
-);
+)((_state, callId) => callId || "");
 
 /**
  * Get machine system_ids in a list request.
@@ -560,25 +557,26 @@ const list = createSelector(
  * @param selected - Whether to filter for selected machines.
  * @returns A list of machines.
  */
-const listIds = (
-  machineState: MachineState,
-  callId: string | null | undefined
-): Machine[MachineMeta.PK][] => {
-  const ids: Machine[MachineMeta.PK][] = [];
-  // Get the ids for all machines in a list or count re
-  const list = getList(machineState, callId);
-  if (list) {
-    // Get the ids for all machines in a list request matching the callId
-    list.groups?.forEach((group) => {
-      group.items.forEach((systemId) => {
-        if (!ids.includes(systemId)) {
-          ids.push(systemId);
-        }
+const listIds = createCachedSelector(
+  machineState,
+  (_state: RootState, callId: string | null | undefined) => callId,
+  (machineState, callId) => {
+    const ids: Machine[MachineMeta.PK][] = [];
+    // Get the ids for all machines in a list or count re
+    const list = getList(machineState, callId);
+    if (list) {
+      // Get the ids for all machines in a list request matching the callId
+      list.groups?.forEach((group) => {
+        group.items.forEach((systemId) => {
+          if (!ids.includes(systemId)) {
+            ids.push(systemId);
+          }
+        });
       });
-    });
+    }
+    return ids;
   }
-  return ids;
-};
+)((_state, callId) => callId || "");
 
 /**
  * Get the ids of machines in a list or details call that are not being used