… src/app/machines/views/MachineList/MachineListTable/StatusColumn/__snapshots__/StatusColumn.test.tsx.snap
… yarn.lock